Move record Edge from the KruskalGenerator
