+++ src/main/java/edu/project2/model/Edge.java
@@ -1,0 +1,4 @@
+package edu.project2.model;
+
+public record Edge(int firstCell, int secondCell) {
+}
